mex-thumbnailer: improve video seeking algorithm
The original algorithm allowed for the thumbnail seek to take at most 3.2s,
which would be failing with some video formats (e.g., mp4) in cases where the
randomly requested seek position is too far into the file. This patch adjusts
the seek algorithm so that (a) the seek position is clamped down to no more
than 90s and (b) the seek is allowed to take up to 1/4 of the requested seek
position.

Signed-off-by: Tomas Frydrych <tomas@sleepfive.com>